+++ src/panelWebView/components/DataBlock/DataBlockSelector.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { useCallback, useEffect, useState } from 'react';
 import { Field, FieldGroup } from '../../../models';
-import { VSCodeDropdown, VSCodeOption } from '@vscode/webview-ui-toolkit/react';
-import { LocalizationKey } from '../../../localization';
+import { LocalizationKey, localize } from '../../../localization';
+import { Dropdown as VSCodeDropdown, DropdownOption } from 'vscrui';
 
 export interface IDataBlockSelectorProps {
   field: Field;
@@ -31,8 +31,8 @@ export const DataBlockSelector: React.FunctionComponent<IDataBlockSelectorProps>
   }, []);
 
   const onGroupSelect = useCallback(
-    (event: any) => {
-      const group = event.currentTarget.value === EMPTY_OPTION ? null : event.currentTarget.value;
+    (selection: DropdownOption) => {
+      const group = selection.value === EMPTY_OPTION ? null : selection.value;
 
       if (group) {
         onGroupSet(group);
@@ -71,23 +71,17 @@ export const DataBlockSelector: React.FunctionComponent<IDataBlockSelectorProps>
 
   return (
     <div className="json_data__selector">
-      <h3>{LocalizationKey.panelDataBlockDataBlockSelectorLabel}</h3>
+      <h3>{localize(LocalizationKey.panelDataBlockDataBlockSelectorLabel)}</h3>
 
       <VSCodeDropdown
         value={selectedGroup ?? EMPTY_OPTION}
-        onChange={onGroupSelect}
-        style={{
-          width: '100%',
-          marginBottom: '1rem'
-        }}
-      >
-        <VSCodeOption value={EMPTY_OPTION}>&nbsp;</VSCodeOption>
-        {options.map((o) => (
-          <VSCodeOption key={o} value={o}>
-            {o}
-          </VSCodeOption>
-        ))}
-      </VSCodeDropdown>
+        onChange={(value) => onGroupSelect(value as DropdownOption)}
+        className='!block mb-4'
+        options={[
+          { value: EMPTY_OPTION, label: '\u00A0' },
+          ...options.map((o) => ({ value: o, label: o }))
+        ]}
+      />
     </div>
   );
 };